move AnimationPlayer into its own files
