+++ apps/freeablo/faworld/actor/basestate.cpp
@@ -61,9 +61,9 @@ namespace FAWorld
             actor.getCurrentFrame(sprite, frame);
 
             if(actor.mPos.mMoving && sprite == actor.mIdleAnim)
-                actor.playAnimation(AnimState::walk, AnimationPlayer::AnimationType::Looped);
+                actor.playAnimation(AnimState::walk, FARender::AnimationPlayer::AnimationType::Looped);
             else if(!actor.mPos.mMoving && sprite == actor.mWalkAnim)
-                actor.playAnimation(AnimState::idle, AnimationPlayer::AnimationType::Looped);
+                actor.playAnimation(AnimState::idle, FARender::AnimationPlayer::AnimationType::Looped);
 
             actor.mPos.update();
